==> 03_Script/01_GlobalHeterogeneity/00B_Analysis_Params.R <==
###############################################################################
# This file defines ANALYSIS parameters as global variables that will be loaded
# before analysis starts. It should define common parameters used by the current
# analysis


#Colors

palette<-c('NK1C'='#F8766D','NKint'='#8494FF',
           'NK1A'='#0CB702',
           'NK1B'='#00BFC4','NK2'='#ED68ED',
           'NK3'='#ABA300')

# Seed for pseudo-random numbers
SEED = 42;


# Scaling parameters (see Seurat::ScaleData())
DATA_CENTER       = TRUE
DATA_SCALE        = TRUE
DATA_VARS_REGRESS = NULL  # c("nCount_RNA") for UMIs (NULL to ignore)

#### Analysis parameters

# Maximum number of variable features to keep
VARIABLE_FEATURES_MAXNB   = 2000;  # For analysis (PCA)
VARIABLE_FEATURES_SHOWTOP = 200;   # For table in report


# Nearest-neighbor graph construction
FINDNEIGHBORS_K = 30

# Cluster identification parameters
FINDCLUSTERS_RESOLUTION     = 0.5;
FINDCLUSTERS_USE_PCA_NBDIMS = 30;  # Number of dimensions to use from PCA results
FINDCLUSTERS_ALGORITHM      = 1;   # 1 = Louvain; 2 = Louvain with multilevel refinement; 3 = SLM; 4 = Leiden
FINDCLUSTERS_DIMS = 1:20

# Parameters for identification of marker annotations for clusters (see Seurat::FindAllMarkers())
FINDMARKERS_METHOD    = "wilcox"  # Method used to identify markers
FINDMARKERS_ONLYPOS   = TRUE     # Only consider overexpressed annotations for markers ? (if FALSE downregulated genes can also be markers)

FINDMARKERS_MINPCT    = 0.2      # Only test genes that are detected in a minimum fraction of cells in either of the two populations. Speed up the function by not testing genes that are very infrequently expressed. Default is '0.1'.
FINDMARKERS_LOGFC_THR = 0.25     # Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is '0.25'. Increasing logfc.threshold speeds up the function, but can miss weaker signals.
FINDMARKERS_PVAL_THR  = 0.001    # PValue threshold for identification of significative markers
FINDMARKERS_SHOWTOP   = 20      # Number of marker genes to show in report and tables (NULL for all)

NUMBER_TOP_SCORING = 20

#Order legends:

ORDER_CLUST_LEGEND = c(  "NK1A" ,  "NK1B", "NK1C", "NKint","NK2" , "NK3")

ORDER_CLUST_LEGEND2 = c(  "NK1A" ,  "NK1B", "NK1C",  "NKint","NK2" , "NK3A", "NK3B", "NK3C")
ORDER_CLUST_LEGEND3 = c("NK1","NK2","NK3")

#Colors

palette<-c('NK1C'='#F8766D','NKint'='#8494FF',
           'NK1A'='#0CB702',
           'NK1B'='#00BFC4','NK2'='#ED68ED',
           'NK3'='#ABA300')



palette2 = c('NK1C'='#F8766D','NKint'='#8494FF',
             'NK1A'='#0CB702',
             'NK1B'='#00BFC4','NK2'='#ED68ED',
             'NK3A'='#C77CFF', 'NK3C'="#ABA300", 'NK3B'="#00A9FF")

palette3 = c('NK1'='#F8766D','NK3'='#0CB702',
             'NK2'='#8494FF')

palette4 = c('NK1'='#F8766D','NK3'='#0CB702',
             'NK2'='#8494FF', 'NKint'='#C77CFF')
